requirements.txt
torch
torchvision
matplotlib
scipy
Pillow
transformers

# Optional: TensorRT backend
# tensorrt
# torch_tensorrt
